+++ src/modules/ngx-typeahead.component.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient } from "@angular/common/http";
 import {
   ChangeDetectorRef,
   Component,
@@ -12,20 +12,36 @@ import {
   TemplateRef,
   ViewChild,
   ViewContainerRef
-} from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
-
-import { Key } from '../models';
+} from "@angular/core";
+import { Observable } from "rxjs/Observable";
+import { Subject } from "rxjs/Subject";
+import { Subscription } from "rxjs/Subscription";
+import { of } from "rxjs/observable/of";
 import {
-  validateNonCharKeyCode,
-  isIndexActive,
+  concat,
+  debounceTime,
+  distinctUntilChanged,
+  filter,
+  map,
+  switchMap,
+  takeUntil
+} from "rxjs/operators";
+import { Key } from "../models";
+import {
   createParamsForQuery,
+  hasCharacters,
+  isEnterKey,
+  isEscapeKey,
+  isIndexActive,
   resolveApiMethod,
+  resolveNextIndex,
+  toFormControlValue,
+  toJsonpFinalResults,
+  toJsonpSingleResult,
   validateArrowKeys,
-  resolveNextIndex
-} from '../services/ngx-typeahead.utils';
-import { Subject } from 'rxjs/Subject';
+  validateNonCharKeyCode
+} from "../services/ngx-typeahead.utils";
+
 /*
  using an external template:
  <input [taItemTpl]="itemTpl" >
@@ -35,13 +51,13 @@ import { Subject } from 'rxjs/Subject';
   </ng-template>
 */
 @Component({
-  selector: '[ngxTypeahead]',
+  selector: "[ngxTypeahead]",
   styles: [
     `
-  .results {
+  .ta-results {
     position: absolute;
   }
-  .typeahead-backdrop {
+  .ta-backdrop {
     bottom: 0;
     left: 0;
     position: fixed;
@@ -49,17 +65,18 @@ import { Subject } from 'rxjs/Subject';
     top: 0;
     z-index: 1;
   }
-  .list-group-item {
+  .ta-item {
     position: relative;
     z-index: 2;
+    display: block;
   }
   `
   ],
   template: `
   <ng-template #suggestionsTplRef>
-  <section class="list-group results" *ngIf="showSuggestions">
-    <div class="typeahead-backdrop" (click)="hideSuggestions()"></div>
-    <button type="button" class="list-group-item"
+  <section class="ta-results list-group" *ngIf="showSuggestions">
+    <div class="ta-backdrop" (click)="hideSuggestions()"></div>
+    <button type="button" class="ta-item list-group-item"
       *ngFor="let result of results; let i = index;"
       [class.active]="markIsActive(i, result)"
       (click)="handleSelectSuggestion(result)">
@@ -78,51 +95,58 @@ export class NgxTypeAheadComponent implements OnInit, OnDestroy {
   results: string[];
 
   @Input() taItemTpl: TemplateRef<any>;
-  @Input() taUrl = '';
+  @Input() taUrl = "";
   @Input() taParams = {};
-  @Input() taQueryParam = 'q';
+  @Input() taQueryParam = "q";
   @Input() taCallbackParamValue;
-  @Input() taApi = 'jsonp';
-  @Input() taApiMethod = 'get';
-  @Input() taResponseTransform: () => void;
+  @Input() taApi = "jsonp";
+  @Input() taApiMethod = "get";
   @Input() taList = [];
+  @Input() taDebounce = 300;
 
   @Output() taSelected = new EventEmitter<string>();
 
-  @ViewChild('suggestionsTplRef') suggestionsTplRef: TemplateRef<any>;
+  @ViewChild("suggestionsTplRef") suggestionsTplRef: TemplateRef<any>;
 
   private suggestionIndex = 0;
   private subscriptions: Subscription[];
   private activeResult: string;
+  private keydown$ = new Subject<KeyboardEvent>();
+  private keyup$ = new Subject<KeyboardEvent>();
 
   constructor(
     private element: ElementRef,
     private viewContainer: ViewContainerRef,
     private http: HttpClient,
     private cdr: ChangeDetectorRef
-  ) { }
+  ) {}
 
-  @HostListener('keydown', ['$event'])
+  @HostListener("keydown", ["$event"])
   handleEsc(event: KeyboardEvent) {
-    if (event.keyCode === Key.Escape) {
+    if (isEscapeKey(event)) {
       this.hideSuggestions();
       event.preventDefault();
     }
+    this.keydown$.next(event);
+  }
+
+  @HostListener("keyup", ["$event"])
+  onkeyup(event: KeyboardEvent) {
+    event.preventDefault();
+    event.stopPropagation();
+    this.keyup$.next(event);
   }
 
   ngOnInit() {
-    const onkeyDown$ = this.onElementKeyDown();
-    this.subscriptions = [
-      this.filterEnterEvent(onkeyDown$),
-      this.listenAndSuggest(),
-      this.navigateWithArrows(onkeyDown$)
-    ];
+    this.filterEnterEvent(this.keydown$);
+    this.listenAndSuggest(this.keyup$);
+    this.navigateWithArrows(this.keydown$);
     this.renderTemplate();
   }
 
   ngOnDestroy() {
-    this.subscriptions.forEach(sub => sub.unsubscribe());
-    this.subscriptions.length = 0;
+    this.keydown$.complete();
+    this.keyup$.complete();
   }
 
   renderTemplate() {
@@ -130,53 +154,53 @@ export class NgxTypeAheadComponent implements OnInit, OnDestroy {
     this.cdr.markForCheck();
   }
 
-  onElementKeyDown() {
-    return Observable.fromEvent(this.element.nativeElement, 'keydown').share();
+  listenAndSuggest(obs: Subject<KeyboardEvent>) {
+    return obs
+      .pipe(
+        filter((e: KeyboardEvent) => validateNonCharKeyCode(e.keyCode)),
+        map(toFormControlValue),
+        debounceTime(this.taDebounce),
+        concat(),
+        distinctUntilChanged(),
+        filter(hasCharacters),
+        switchMap((query: string) => this.suggest(query))
+      )
+      .subscribe((results: string[]) => {
+        this.results = results;
+        this.displaySuggestions(Key.ArrowDown);
+      });
   }
 
-  filterEnterEvent(elementObs: Observable<{}>) {
+  filterEnterEvent(elementObs: Subject<KeyboardEvent>) {
     return elementObs
-      .filter((e: KeyboardEvent) => e.keyCode === Key.Enter)
-      .subscribe((event: Event) => {
-        event.preventDefault();
+      .pipe(filter(isEnterKey))
+      .subscribe((event: KeyboardEvent) => {
         this.handleSelectSuggestion(this.activeResult);
       });
   }
 
-  listenAndSuggest() {
-    return Observable.fromEvent(this.element.nativeElement, 'keyup')
-      .filter((e: any) => validateNonCharKeyCode(e.keyCode))
-      .map((e: any) => e.target.value)
-      .debounceTime(300)
-      .concat()
-      .distinctUntilChanged()
-      .filter((query: string) => query.length > 0)
-      .switchMap((query: string) => this.suggest(query))
-      .subscribe((results: string[]) => {
-        this.results = results;
-        this.showSuggestions = true;
-        this.suggestionIndex = 0;
-        this.cdr.markForCheck();
-      });
+  navigateWithArrows(elementObs: Subject<KeyboardEvent>) {
+    return elementObs
+      .pipe(
+        filter((e: any) => validateArrowKeys(e.keyCode)),
+        map((e: any) => e.keyCode)
+      )
+      .subscribe((keyCode: number) => this.displaySuggestions(keyCode));
   }
 
-  navigateWithArrows(elementObs: Observable<{}>) {
-    return elementObs
-      .filter((e: any) => validateArrowKeys(e.keyCode))
-      .map((e: any) => e.keyCode)
-      .subscribe((keyCode: number) => {
-        this.suggestionIndex = resolveNextIndex(
-          this.suggestionIndex,
-          keyCode === Key.ArrowDown
-        );
-        this.showSuggestions = true;
-        this.cdr.markForCheck();
-      });
+  displaySuggestions(keyCode: number) {
+    this.suggestionIndex = resolveNextIndex(
+      this.suggestionIndex,
+      keyCode === Key.ArrowDown,
+      this.results.length
+    );
+    this.showSuggestions = true;
+    this.cdr.markForCheck();
   }
 
   suggest(query: string) {
     return this.taList.length
-      ? this.createListSource(this.taList)
+      ? this.createListSource(this.taList, query)
       : this.request(query);
   }
 
@@ -194,7 +218,7 @@ export class NgxTypeAheadComponent implements OnInit, OnDestroy {
     const options = {
       params: searchConfig
     };
-    const isJsonpApi = this.taApi === 'jsonp';
+    const isJsonpApi = this.taApi === "jsonp";
     return isJsonpApi
       ? this.requestJsonp(url, options, this.taCallbackParamValue)
       : this.requestHttp(url, options);
@@ -202,16 +226,14 @@ export class NgxTypeAheadComponent implements OnInit, OnDestroy {
 
   requestHttp(url: string, options) {
     const apiMethod = resolveApiMethod(this.taApiMethod);
-    const responseTransformMethod = this.taResponseTransform || ((item: {}) => item);
-    return this.http[apiMethod](url, options)
-      .map((results: any[]) => results.map(responseTransformMethod));
+    return this.http[apiMethod](url, options);
   }
 
-  requestJsonp(url, options, callback = 'callback') {
+  requestJsonp(url, options, callback = "callback") {
     const params = options.params.toString();
-    return this.http.jsonp(`${url}?${params}`, callback)
-      .map((response: Response) => response[1])
-      .map((results: any[]) => results.map((result: string) => result[0]));
+    return this.http
+      .jsonp(`${url}?${params}`, callback)
+      .pipe(map(toJsonpSingleResult), map(toJsonpFinalResults));
   }
 
   markIsActive(index: number, result: string) {
@@ -235,9 +257,7 @@ export class NgxTypeAheadComponent implements OnInit, OnDestroy {
     return this.taItemTpl !== undefined;
   }
 
-  createListSource(list: any[]): Observable<string[]> {
-    return new Observable((observer) => {
-      observer.next(list);
-    });
+  createListSource(list: any[], query: string): Observable<string[]> {
+    return of(list.filter((item: string) => item.includes(query)));
   }
 }