+++ tests/modules/ngx-typeahead.component.spec.ts
@@ -1,20 +1,16 @@
-import {
-  TestBed,
-  ComponentFixture,
-  inject
-} from '@angular/core/testing';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { TestBed, ComponentFixture, inject } from "@angular/core/testing";
+import { HttpClientModule, HttpClient } from "@angular/common/http";
 import {
   ChangeDetectorRef,
   ElementRef,
   ViewContainerRef,
   DebugElement
-} from '@angular/core';
-import { NgxTypeAheadComponent } from '../../src/modules/ngx-typeahead.component';
-import { NgxTypeaheadModule } from '../../src/modules/ngx-typeahead.module';
-import { Key } from '../../src/models';
+} from "@angular/core";
+import { NgxTypeAheadComponent } from "../../src/modules/ngx-typeahead.component";
+import { NgxTypeaheadModule } from "../../src/modules/ngx-typeahead.module";
+import { Key } from "../../src/models";
 
-describe('A Typeahead component', () => {
+describe("A Typeahead component", () => {
   let component: NgxTypeAheadComponent;
   let fixture: ComponentFixture<NgxTypeAheadComponent>;
   // let de: DebugElement;
@@ -26,17 +22,21 @@ describe('A Typeahead component', () => {
   // register all needed dependencies
   beforeEach(() => {
     // setting mocked providers
-    spyElementRef = jasmine.createSpyObj('spyElementRef', ['nativeElement']);
-    spyViewContainerRef = jasmine.createSpyObj('spyViewContainerRef', ['createEmbeddedView']);
-    spyHttp = jasmine.createSpyObj('spyHttp', ['get']);
-    spyChangeDetectorRef = jasmine.createSpyObj('spyChangeDetectorRef', ['markForCheck']);
+    spyElementRef = jasmine.createSpyObj("spyElementRef", ["nativeElement"]);
+    spyViewContainerRef = jasmine.createSpyObj("spyViewContainerRef", [
+      "createEmbeddedView"
+    ]);
+    spyHttp = jasmine.createSpyObj("spyHttp", ["get", "post"]);
+    spyChangeDetectorRef = jasmine.createSpyObj("spyChangeDetectorRef", [
+      "markForCheck"
+    ]);
 
     // setting spy on methods
-    spyOn(NgxTypeAheadComponent.prototype, 'filterEnterEvent').and.callThrough();
-    spyOn(NgxTypeAheadComponent.prototype, 'listenAndSuggest').and.callThrough();
-    spyOn(NgxTypeAheadComponent.prototype, 'navigateWithArrows').and.callThrough();
-    spyOn(NgxTypeAheadComponent.prototype, 'onElementKeyDown').and.callThrough();
-    spyOn(NgxTypeAheadComponent.prototype, 'renderTemplate');
+    ["filterEnterEvent", "listenAndSuggest", "navigateWithArrows"].forEach(
+      method =>
+        spyOn(NgxTypeAheadComponent.prototype, method as any).and.callThrough()
+    );
+    spyOn(NgxTypeAheadComponent.prototype, "renderTemplate");
 
     TestBed.configureTestingModule({
       declarations: [NgxTypeAheadComponent],
@@ -45,7 +45,7 @@ describe('A Typeahead component', () => {
         { provide: ElementRef, useValue: spyElementRef },
         { provide: ViewContainerRef, useValue: spyViewContainerRef },
         { provide: HttpClient, useValue: spyHttp },
-        { provide: ChangeDetectorRef, useValue: spyChangeDetectorRef },
+        { provide: ChangeDetectorRef, useValue: spyChangeDetectorRef }
       ]
     });
 
@@ -54,60 +54,127 @@ describe('A Typeahead component', () => {
     fixture.detectChanges();
   });
 
-  it('should have an instance', () => {
+  it("should have an instance", () => {
     expect(component).toBeDefined();
   });
 
-  describe('Default Inputs', () => {
-    it('should have a default taQueryParam', () => {
-      expect(component.taQueryParam).toBe('q');
+  describe("Default Inputs", () => {
+    it("should have a default taQueryParam", () => {
+      expect(component.taQueryParam).toBe("q");
     });
 
-    it('should have a default undefined taCallbackParamValue', () => {
+    it("should have a default undefined taCallbackParamValue", () => {
       expect(component.taCallbackParamValue).toBeUndefined();
     });
 
-    it('should have a default taApi', () => {
-      expect(component.taApi).toBe('jsonp');
+    it("should have a default taApi", () => {
+      expect(component.taApi).toBe("jsonp");
     });
 
-    it('should have a default taApiMethod', () => {
-      expect(component.taApiMethod).toBe('get');
+    it("should have a default taApiMethod", () => {
+      expect(component.taApiMethod).toBe("get");
     });
   });
 
-  describe('Component Init', () => {
-    it('should create a keydown observable', () => {
-      expect(component.onElementKeyDown).toHaveBeenCalled();
-    });
-
-    it('should create a filter enter event subscription', () => {
+  describe("Component Init", () => {
+    it("should create a filter enter event subscription", () => {
       expect(component.filterEnterEvent).toHaveBeenCalledTimes(1);
     });
 
-    it('should create a listen and suggest subscription', () => {
+    it("should create a listen and suggest subscription", () => {
       expect(component.listenAndSuggest).toHaveBeenCalledTimes(1);
     });
 
-    it('should create a navigation with arrows subscription', () => {
+    it("should create a navigation with arrows subscription", () => {
       expect(component.navigateWithArrows).toHaveBeenCalledTimes(1);
     });
 
-    it('should render template', () => {
+    it("should render template", () => {
       expect(component.renderTemplate).toHaveBeenCalled();
     });
   });
 
-  describe('Functionality', () => {
-    it('should hide suggestion when ESC is pressed', () => {
+  describe("Functionality", () => {
+    it("should hide suggestion when ESC is pressed", () => {
       const mockedEvent = {
         keyCode: Key.Escape,
-        preventDefault: () => undefined,
+        preventDefault: () => undefined
       } as KeyboardEvent;
       component.handleEsc(mockedEvent);
       const actual = component.showSuggestions;
       const expected = false;
       expect(actual).toBe(expected);
+    });
+  });
+
+  describe("Requests", () => {
+    [
+      {
+        title: "should request with http get",
+        method: "get",
+        inputs: {
+          taUrl: "mocked",
+          taApi: "http"
+        },
+        expected: "get"
+      },
+      {
+        title: "should request with http post",
+        method: "post",
+        inputs: {
+          taUrl: "mocked",
+          taApi: "http"
+        },
+        expected: "post"
+      }
+    ].forEach(({ title, method, inputs, expected }) => {
+      it(title, () => {
+        const query = "something";
+        component.taUrl = inputs.taUrl;
+        component.taApi = inputs.taApi;
+        component.taApiMethod = method;
+        component.suggest(query);
+        expect(spyHttp[expected]).toHaveBeenCalled();
+      });
+    });
+
+    it("should show a static list when given", () => {
+      component.taList = ["aa", "ab", "ac"];
+      const query = "a";
+      const actual = component.suggest(query);
+      actual.subscribe(list => {
+        expect(list.join("")).toMatch(component.taList.join(""));
+      });
+    });
+
+    it("should show a filtered static list when given", () => {
+      component.taList = ["aa", "bb", "cc"];
+      const query = "a";
+      const actual = component.suggest(query);
+      actual.subscribe(list => {
+        expect(list.join("")).toMatch(component.taList[0]);
+      });
+    });
+
+    it("should show an empty static list when query not exist in results", () => {
+      component.taList = ["aa", "bb", "cc"];
+      const query = "x";
+      const actual = component.suggest(query);
+      actual.subscribe(list => {
+        expect(list.length).toBe(0);
+      });
+    });
+  });
+
+  describe("Subscriptions", () => {
+    it("should unsubscribe from all subscriptions", () => {
+      spyOn(component, "handleSelectSuggestion");
+      component.ngOnDestroy();
+      const event = {
+        keyCode: Key.Enter
+      } as KeyboardEvent;
+      component.handleEsc(event);
+      expect(component.handleSelectSuggestion).not.toHaveBeenCalled();
     });
   });
 });